feat: Use semantic-release for tags
Up until now, auto-build-image was released under master:stable and that
tag was being overwritten on each commit to master.

This was fine as long as the image remained backwards-compatible, but
this is not viable in the long run: Eventually, just upgrading the
docker base image is bound to break things for people on older versions
of GitLab.

Resolves https://gitlab.com/gitlab-org/cluster-integration/auto-build-image/-/issues/11